refactor : 보너스볼 추가 및 display를 prize로

+++ src/main/java/lotto/domain/Prize.java
@@ -37,7 +37,7 @@ public enum Prize {
                 .collect(Collectors.toList());
     }
 
-    public long calculatePriceMoney(long count) {
+    public int calculatePriceMoney(int count) {
         return this.prizeMoney * count;
     }
